requirements.txt
asgiref==3.3.4
attrs==21.2.0
boto3==1.17.44
botocore==1.20.72
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-sendgrid-v5==0.9.1
django-storages==1.11.1
django-templated-mail==1.1.1
django-username-email==2.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.1.2
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.7
sendgrid==6.6.0
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.26.4
